feat(leadspace-interior): add spacerLine component. move eyebrow to components/common directory.
