+++ src/components/common/spacerLine/index.ts
@@ -1,0 +1,1 @@
+export { SpacerLine } from './spacerLine';
