+++ src/components/common/spacerLine/spacerLine.scss
@@ -1,0 +1,6 @@
+.kd-spacer-line {
+	background: var(--kd-color-background-accent-primary);
+	height: 2px;
+	margin: var(--kd-spacing-8) 0;
+	width: 16px;
+}
