+++ src/components/leadspaces/leadspaceInterior/leadspaceInterior.scss
@@ -1,4 +1,5 @@
 @use '../../../scss/utility/component-palettes.scss';
+@use '~@kyndryl-design-system/shidoka-foundation/scss/reset.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/variables/breakpoints.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/grid.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/typography.scss';
