+++ src/index.ts
@@ -1,4 +1,5 @@
-export { Eyebrow } from './components/eyebrow';
+export { Eyebrow } from './components/common/eyebrow';
+export { SpacerLine } from './components/common/spacerLine';
 export { AlternatingFeatureWithImage } from './components/alternatingFeatureWithImage';
 export { StatsCarousel, StatsCarouselSlide } from './components/patterns/statsCarousel';
 export { PersistentCta } from './components/callToAction/persistentCta';
